Removed using of query clauses
Also the TP country management is fully on utils.py. The reviews function on site.py has been removed because it was just a wrapper of extract_reviews of _xray.py. Moreover, the first request that was used just to know the number of pages of a result listing page now it is used also to extract company information and reviews. This reduce the number of requests by one for each search request.