Added tox conf and reimplemented CompnayDoc
The class CompanyDoc has been reimplemented by removing the self.attrs dictionary and adding the decorators @property and @cached_property. Also the xray methods now manages the exceptions when Trustpilot forbids the access to its data. The tests methods for xray are now based on local files to reduce the number of accesses to Trustpilot.