Better modularity
All the methods not related to the extraction of information from the HTML page has been transfered to the site module. Now the extraction of information from a local file works. Instead of getting just the reviews from a source, which is something unlikely to do, the method get_company let the user get the information of just one company, including its reviews